+++ packages/react-universal/src/hooks/useFinishNavigation.ts
@@ -1,0 +1,12 @@
+import React from "react";
+
+import useResponse from "./useResponse";
+
+export default function useFinishNavigation() {
+  const { navigation } = useResponse();
+  React.useEffect(() => {
+    if (navigation.finish) {
+      navigation.finish();
+    }
+  }, [navigation]);
+}