+++ packages/types/src/index.ts
@@ -14,6 +14,7 @@ export interface CuriRouter {
   url(details: RouteLocation): string;
   navigate(options: NavigationDetails): (() => void) | void;
   destroy(): void;
+  suspends(): boolean;
   route: RouteGetter;
   history: History;
   external: any;
@@ -39,6 +40,7 @@ export interface CurrentResponse {
 export interface Navigation {
   action: Action;
   previous: Response | undefined;
+  finish?: () => void;
 }
 
 // configuration options for router.observe and router.once